+++ src/index.ts
@@ -1,3 +1,0 @@
-export { NgrxUtilsModule } from './module';
-export { ofAction } from './of-action';
-export { Select, NgrxSelect } from './select';