+++ src/public_api.ts
@@ -1,0 +1,3 @@
+export { NgrxUtilsModule } from './module';
+export { ofAction } from './of-action';
+export { Select, NgrxSelect } from './select';